+++ view/frontend/requirejs-config.js
@@ -1,7 +1,8 @@
 var config = {
     map: {
         '*': {
-            toosoClickAfterSearch: 'Bitbull_Tooso/js/clickAfterSearch'
+            toosoClickAfterSearch: 'Bitbull_Tooso/js/clickAfterSearch',
+            toosoSpeechToText: 'Bitbull_Tooso/js/speechToText'
         }
     }
 };