+++ view/frontend/web/js/clickAfterSearch.js
@@ -1,33 +1,48 @@
-/* global varienGlobalEvents, ta */
+/* global ta */
 
 define([
-    "domReady!",
-], function() {
+    'jquery',
+], function($) {
 
     /**
-     * Elements
+     * Global settings
+     *
+     * @type {*}
+     */
+    var config = {};
+
+    /**
+     * Product links elements
+     *
      * @type {Array}
      */
     var elements = [];
 
     /**
-     * Configuration function
-     *
-     * @param {Object} config
-     * @param {*} element
+     * Find alla products links into container
      */
-    return function (config, element) {
-        var linkValue = element.getAttribute('href');
-        var productSku = element.getAttribute(config.attributeName);
+    function elaborateProductsElements(parentContainer) {
+        elements = [];
+        $(parentContainer).find(config.productLinksSelector).each(function () {
+            var element = $(this);
+            elements.push(element);
+            var productSku = element.attr(config.attributeName);
+            var linkValue = element.attr('href');
+            element.attr('data-href', linkValue);
+            element.attr('href', '#');
+            attachClickHandler(element, productSku, linkValue);
+        });
+    }
 
-        config.pageSize = config.pageSize || 0;
-        config.currentPage = config.currentPage || 1;
-
-        elements.push(element);
-
-        element.setAttribute('data-href', linkValue);
-        element.setAttribute('href', '#');
-        element.addEventListener('click', function () {
+    /**
+     * Attach click handler to element
+     *
+     * @param {*} element
+     * @param {String} productSku
+     * @param {String} linkValue
+     */
+    function attachClickHandler(element, productSku, linkValue) {
+        element.click(function () {
             if (window.ta === undefined) {
                 console.warn('Tooso: ta is not include but analytics is active');
                 document.location.href = linkValue; // this should never happens
@@ -64,6 +79,70 @@ define([
                     document.location.href = linkValue;
                 },
             });
+        })
+    }
+
+    /**
+     * Observer for HTML changes
+     *
+     * @param {*} target
+     * @param {function} callback
+     */
+    function observeForChanges(target, callback) {
+        var observer = new MutationObserver(callback);
+        observer.observe(target, { attributes: true, childList: true, characterData: true });
+    }
+
+    /**
+     * Elaborate pagination from URL query parameters
+     *
+     */
+    function elaboratePaginationFromURL() {
+        var queries = {};
+        $.each(document.location.search.substr(1).split('&'),function(c,q){
+            var i = q.split('=');
+            queries[i[0].toString()] = i[1].toString();
         });
+        if (queries.product_list_limit !== undefined) {
+            config.pageSize = parseInt(queries.product_list_limit);
+            if (config.debug) {
+                console.debug('Tooso: page size is now '+config.pageSize);
+            }
+        }
+        if (queries.p !== undefined) {
+            config.currentPage = parseInt(queries.p);
+            if (config.debug) {
+                console.debug('Tooso: current page is now '+config.currentPage);
+            }
+        }
+    }
+
+    /**
+     * Configuration function
+     *
+     * @param {Object} config
+     * @param {*} element
+     */
+    return function (configParams, parentContainer) {
+        config = Object.assign({}, configParams);
+        config.pageSize = config.pageSize || 0;
+        config.currentPage = config.currentPage || 1;
+
+        // Elaborate child elements
+        elaborateProductsElements(parentContainer);
+
+        // Observer for HTML changes (like AJAX pagination,filters..)
+        observeForChanges(parentContainer, function () {
+            if (config.debug) {
+                console.debug('Tooso: products container changed');
+            }
+
+            // Check for pagination change
+            elaboratePaginationFromURL();
+
+            // Re-elaborate child elements
+            elaborateProductsElements(parentContainer);
+        })
+
     }
 });