+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-03-27
+### Fixed
+- Fix TA initialisation before inclusion
+### Added
+- Add Javascript SDK
+- Add SpeechToText functionality
+- Add cart product's quantity tracking
+- Add AJAX support for click after search tracking
+- Add product attributes export functionality
+
 ## [1.1.1] - 2019-03-19
 ### Fixed
 - Fix single entity catalog reindex